# #451 remove tail recursion and faster environment records